.github/scripts/check_root_domain_literals/requirements.txt
pyyaml==6.0.2
tldextract==5.1.3
